LPD-45733 Move export tests to new file
