+++ modules/test/playwright/tests/export-import-web/import.spec.ts
@@ -12,7 +12,6 @@ import fs from 'fs/promises';
 import * as path from 'path';
 import {getComparator} from 'playwright-core/lib/utils';
 
-import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
 import {applicationsMenuPageTest} from '../../fixtures/applicationsMenuPageTest';
 import {dataApiHelpersTest} from '../../fixtures/dataApiHelpersTest';
 import {depotAdminPageTest} from '../../fixtures/depotAdminPageTest';
@@ -22,10 +21,7 @@ import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
 import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
 import {pageTemplatesPagesTest} from '../../fixtures/pageTemplatesPagesTest';
-import {pagesAdminPagesTest} from '../../fixtures/pagesAdminPagesTest';
-import {productMenuPageTest} from '../../fixtures/productMenuPageTest';
 import {wikiPagesTest} from '../../fixtures/wikiPagesTest';
-import {depotsPagesTest} from '../../tests/depot-web/fixtures/depotsPagesTest';
 import getRandomString from '../../utils/getRandomString';
 import {getTempDir} from '../../utils/temp';
 import {readFileFromZip} from '../../utils/zip';
@@ -33,25 +29,21 @@ import {exportImportPagesTest} from './fixtures/exportImportPagesTest';
 import {stagingPageTest} from './fixtures/stagingPageTest';
 
 export const test = mergeTests(
-	apiHelpersTest,
 	applicationsMenuPageTest,
 	dataApiHelpersTest,
+	depotAdminPageTest,
 	documentLibraryPagesTest,
+	exportImportPagesTest,
 	featureFlagsTest({
 		'LPD-35013': {enabled: true},
 		'LPD-35914': {enabled: true, system: true},
 	}),
-	productMenuPageTest,
-	exportImportPagesTest,
-	stagingPageTest,
-	depotsPagesTest,
-	depotAdminPageTest,
-	pagesAdminPagesTest,
+	isolatedSiteTest,
+	loginTest(),
 	pageEditorPagesTest,
 	pageTemplatesPagesTest,
-	wikiPagesTest,
-	loginTest(),
-	isolatedSiteTest
+	stagingPageTest,
+	wikiPagesTest
 );
 
 async function getSiteHomePageScreenshot(
@@ -397,129 +389,4 @@ test('can export and import custom object entries at instance level', async ({
 			name: objectEntry.name,
 		})
 	);
-});
-
-test('cannot export site scoped custom object entries at instance level', async ({
-	apiHelpers,
-	applicationsMenuPage,
-	page,
-}) => {
-	const objectActionApiClient =
-		await apiHelpers.buildRestClient(ObjectDefinitionApi);
-
-	const {body: objectDefinition} =
-		await objectActionApiClient.postObjectDefinition({
-			active: true,
-			externalReferenceCode: 'test',
-			label: {
-				en_US: 'Test',
-			},
-			name: 'Test',
-			objectFields: [
-				{
-					DBType: ObjectField.DBTypeEnum.String,
-					businessType: ObjectField.BusinessTypeEnum.Text,
-					indexed: true,
-					indexedAsKeyword: true,
-					label: {
-						en_US: 'Name',
-					},
-					name: 'name',
-					required: true,
-				},
-			],
-			pluralLabel: {
-				en_US: 'Tests',
-			},
-			portlet: true,
-			scope: 'site',
-			status: {
-				code: 0,
-			},
-		});
-
-	apiHelpers.data.push({id: objectDefinition.id, type: 'objectDefinition'});
-
-	await apiHelpers.objectEntry.postObjectEntry(
-		{externalReferenceCode: '', name: 'test'},
-		'c/tests/scopes/Guest'
-	);
-
-	await applicationsMenuPage.goToExport();
-
-	await page.getByTestId('creationMenuNewButton').nth(1).click();
-
-	await expect(page.getByLabel('Tests 1 Items')).toBeHidden();
-});
-
-test('can export custom object entries at instance level with permissions based on selection', async ({
-	apiHelpers,
-	applicationsMenuPage,
-	exportImportPage,
-	page,
-}) => {
-	const objectActionApiClient =
-		await apiHelpers.buildRestClient(ObjectDefinitionApi);
-
-	const {body: objectDefinition} =
-		await objectActionApiClient.postObjectDefinition({
-			active: true,
-			externalReferenceCode: 'test',
-			label: {
-				en_US: 'Test',
-			},
-			name: 'Test',
-			objectFields: [
-				{
-					DBType: ObjectField.DBTypeEnum.String,
-					businessType: ObjectField.BusinessTypeEnum.Text,
-					indexed: true,
-					indexedAsKeyword: true,
-					label: {
-						en_US: 'Name',
-					},
-					name: 'name',
-					required: true,
-				},
-			],
-			pluralLabel: {
-				en_US: 'Tests',
-			},
-			portlet: true,
-			scope: 'company',
-			status: {
-				code: 0,
-			},
-		});
-
-	apiHelpers.data.push({id: objectDefinition.id, type: 'objectDefinition'});
-
-	await apiHelpers.objectEntry.postObjectEntry(
-		{externalReferenceCode: '', name: 'test'},
-		'c/tests'
-	);
-
-	await applicationsMenuPage.goToExport();
-
-	await page.getByTestId('creationMenuNewButton').nth(1).click();
-
-	await page.getByLabel('Tests 1 Items').click();
-
-	const exportName = 'CustomObject-WithPermissions-' + getRandomString();
-
-	await exportImportPage.title.fill(exportName);
-
-	await page.getByLabel('Export Permissions').click();
-
-	await exportImportPage.exportButton.click();
-
-	const exportFilePath =
-		await exportImportPage.downloadExportProcess(exportName);
-
-	const content = await readFileFromZip('C_Test.json', exportFilePath);
-
-	const json = JSON.parse(content);
-
-	expect(json.length).toBe(1);
-	expect(json[0]).toHaveProperty('permissions');
 });
